refactor(routes): support static export routes (#182)
