+++ apps/web/next.config.ts
@@ -1,7 +1,8 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  output: "standalone",
+  output: "export",
+  trailingSlash: true,
   images: {
     dangerouslyAllowSVG: true,
     unoptimized: true,
