--- apps/web/src/app/(home)/[...rest]/page.tsx
@@ -1,5 +1,0 @@
-import { notFound } from "next/navigation";
-
-export default function CatchAllPage() {
-  notFound();
-}
