+++ apps/web/src/app/(home)/post/page.tsx
@@ -1,17 +1,11 @@
-import Image from "next/image";
 import { Suspense } from "react";
-import Balancer from "react-wrap-balancer";
 
-import FilterList from "@/components/filter/filter-list";
-import FilterSelectBox from "@/components/filter/filter-select-box";
 import Loading from "@/components/loading";
-import MarkdownRenderer from "@/components/markdown/markdown-renderer";
 import PageHeader from "@/components/page-header";
-import Pagination from "@/components/pagination";
-import { ProgressBarLink } from "@/components/progress-bar";
 import config from "@/config";
-import { POSTS_PER_PAGE } from "@/lib/constants";
 import { getBlogPosts } from "@/lib/db/v1/post";
+
+import PostListClient from "./post-list-client";
 
 const { title, about } = config;
 
@@ -20,115 +14,16 @@ export const metadata = {
   description: "Read my thoughts on software development, design, and more.",
 };
 
-type BlogQueryParams = Promise<{ tag?: string; page?: string }>;
-
-async function BlogPosts({ searchParams }: { searchParams: BlogQueryParams }) {
-  const { tag = "All", page = "1" } = await searchParams;
+export default async function Post() {
   const allBlogs = await getBlogPosts();
-  const blogTags = [
-    "All",
-    ...Array.from(
-      new Set(allBlogs.map((post) => post.metadata.category ?? ""))
-    ),
-  ];
-  const selectedTag = tag;
-  const currentPage = parseInt(page, 10);
+  const posts = allBlogs.map(({ slug, metadata }) => ({ slug, metadata }));
 
-  // Filter blogs based on the selected tag
-  const filteredBlogs =
-    selectedTag === "All"
-      ? allBlogs
-      : allBlogs.filter((post) => post.metadata.category === selectedTag);
-
-  // Calculate total pages
-  const totalPages = Math.ceil(filteredBlogs.length / POSTS_PER_PAGE);
-
-  // Get blogs for current page
-  const paginatedBlogs = filteredBlogs.slice(
-    (currentPage - 1) * POSTS_PER_PAGE,
-    currentPage * POSTS_PER_PAGE
-  );
-
-  return (
-    <section className="blog-posts">
-      <FilterList path="post" selectedTag={selectedTag} blogTags={blogTags} />
-      <FilterSelectBox
-        path="post"
-        selectedTag={selectedTag}
-        blogTags={blogTags}
-      />
-      <ul className="blog-posts-list">
-        {paginatedBlogs.map((post) => (
-          <li
-            key={post.slug}
-            className="blog-post-item active"
-            data-category={post.metadata.category}
-          >
-            <ProgressBarLink
-              href={`/post/${post.slug}`}
-              rel="noopener noreferrer"
-            >
-              <figure className="blog-banner-box">
-                <Image
-                  src={post.metadata.banner}
-                  alt={post.metadata.alt || "Blog post image"}
-                  width={1600}
-                  height={900}
-                  priority={false}
-                  placeholder="blur"
-                  loading="eager"
-                  blurDataURL="https://docs.1chooo.com/images/cover-with-1chooo-com.png"
-                />
-              </figure>
-              <div className="blog-content">
-                <div className="blog-meta">
-                  <p className="blog-category">{post.metadata.category}</p>
-                  <span className="dot"></span>
-                  <time dateTime={post.metadata.publishedAt}>
-                    {new Date(post.metadata.publishedAt).toLocaleDateString(
-                      "en-us",
-                      {
-                        month: "long",
-                        day: "numeric",
-                        year: "numeric",
-                      }
-                    )}
-                  </time>
-                </div>
-                <h3 className="text-white-2 hover:text-orange-yellow-crayola text-2xl font-semibold leading-[1.3] transition-all">
-                  <Balancer>
-                    <MarkdownRenderer content={post.metadata.title} />
-                  </Balancer>
-                </h3>
-                <div className="text-light-gray text-s line-clamp-2 overflow-hidden font-light leading-6">
-                  <MarkdownRenderer content={post.metadata.summary} />
-                </div>
-              </div>
-            </ProgressBarLink>
-          </li>
-        ))}
-      </ul>
-      <Pagination
-        currentPage={currentPage}
-        totalPages={totalPages}
-        selectedTag={selectedTag}
-        basePath="/post"
-      />
-    </section>
-  );
-}
-
-export default function Post({
-  searchParams,
-}: {
-  searchParams: BlogQueryParams;
-}) {
   return (
     <article>
       <PageHeader header={`${about.preferredName}'s Blog`} />
 
       <Suspense fallback={<Loading />}>
-        <BlogPosts searchParams={searchParams} />
+        <PostListClient posts={posts} />
       </Suspense>
     </article>
   );
