--- apps/web/src/app/(metadata)/manifest.webmanifest/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function Manifest() {
-  redirect("/manifest.json");
-}
